#5 [refactor] common btn 폴더링 변경
