+++ feature/onboarding/src/main/java/com/sopetit/onboarding/dollnaming/DollNamingScreen.kt
@@ -52,7 +52,7 @@ import com.sopetit.design_system.Gray700
 import com.sopetit.design_system.R
 import com.sopetit.design_system.Softie
 import com.sopetit.design_system.SoftieTypo
-import com.sopetit.ui.common.item.BottomRectangleBtn
+import com.sopetit.ui.common.button.BottomRectangleBtn
 import com.sopetit.ui.common.topbar.OnboardingTopBar
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
