+++ feature/onboarding/src/main/java/com/sopetit/onboarding/dolltype/DollTypeChoiceScreen.kt
@@ -32,7 +32,7 @@ import com.sopetit.design_system.Gray700
 import com.sopetit.design_system.SoftieTypo
 import com.sopetit.domain.entity.enums.DollType
 import com.sopetit.onboarding.model.DollTypeModel
-import com.sopetit.ui.common.item.BottomRectangleBtn
+import com.sopetit.ui.common.button.BottomRectangleBtn
 import com.sopetit.ui.common.topbar.OnboardingTopBar
 
 @Composable
